--- app/views/search/search_filer.html.haml
@@ -1,2 +1,0 @@
-%h1 Search Location #search_location
-%p Find me in app/views/search/search_location.html.haml
