--- app/views/search/search_filer_results.html.haml
@@ -1,25 +1,0 @@
-= render 'location_error'
-= render 'shared/navbar'
-:javascript
-  checkLocationError(#{@location_error.to_json.html_safe})
-- unless @location_error[:error]
-  :javascript
-    initalCenter = #{[@result_lat, @result_lng]}
-    markerData = #{@markers.to_json.html_safe}
-    bounds = #{@bounds}
-
-
-.row
-  .col-xs-8
-    .panel.panel-default
-      .panel-heading
-        Select a marker to see the description
-      .panel-body.panel-map
-        #map
-  .col-xs-4
-    = render 'shared/sidebar'
-
-- unless @location_error[:error]
-  :javascript
-    map = L.map('map').setView(initalCenter, 15)
-    loadMarkers()
